Refatoração e separação dos botões de Ida e Volta
